Various changes
terrain, ghost bounces, etc

+++ Assets/TrackGeneration/Track Parameters.cs
@@ -7,13 +7,13 @@ namespace Assets.TrackGeneration
         public float TrackWidth { get; }
         public float WallHeight { get; }
         public float WallDepth { get; }
-        public int SegmentsPerUnit { get; }
+        public float SegmentsPerUnit { get; }
         public float BankingAngle { get; }
 
         public float SupportCount { get; }
         public const float TRACK_HEIGHT_OFFSET = 0.1f;
 
-        public TrackParameters(float width, float height, float depth, int segments, float banking, float supportCount)
+        public TrackParameters(float width, float height, float depth, float segments, float banking, float supportCount)
         {
             TrackWidth = width;
             WallHeight = height;
